exception handling

+++ src/todos/todos.service.ts
@@ -1,4 +1,4 @@
-import { Injectable, Logger, NotFoundException } from '@nestjs/common';
+import { BadRequestException, HttpStatus, Injectable, Logger, NotFoundException } from '@nestjs/common';
 import { CreateTodoDto } from './dto/create-todo.dto';
 import { UpdateTodoDto } from './dto/update-todo.dto';
 import { InjectModel } from '@nestjs/mongoose';
@@ -14,7 +14,11 @@ export class TodosService {
   async create(createTodoDto: CreateTodoDto): Promise<Todo> {
     const createdTodo = new this.todoModel(createTodoDto);
     this.logger.log(`create todo with title: ${createTodoDto.title}`);
-    return createdTodo.save();
+    const todo = await createdTodo.save();
+    if (!todo) {
+      throw new BadRequestException('Todo not created');
+    }
+    return todo;
   }
 
   async findAll(limit: number, skip: number): Promise<Todo[]> {
@@ -24,10 +28,14 @@ export class TodosService {
   }
 
   async findOne(id: string): Promise<Todo> {
+    const idIsValid = Types.ObjectId.isValid(id);
+    if (!idIsValid) {
+      throw new BadRequestException("Invalid id !");
+    }
     const todo = await this.todoModel.findOne({ _id: id }).exec();
     this.logger.log(todo);
     if (!todo) {
-      throw new Error("Todo not found !");
+      throw new NotFoundException("Todo not found !");
     }
     this.logger.log(`find todo with id: ${id}`);
 
@@ -35,6 +43,10 @@ export class TodosService {
   }
 
   async update(id: string, updateTodoDto: UpdateTodoDto) {
+    const idIsValid = Types.ObjectId.isValid(id);
+    if (!idIsValid) {
+      throw new BadRequestException("Invalid id !");
+    }
     const todo = await this.todoModel.updateOne({ _id: id }, updateTodoDto).exec();
     if (todo.modifiedCount === 0) {
       throw new Error("Todo not found !");
@@ -43,6 +55,10 @@ export class TodosService {
   }
 
   async remove(id: string) {
+    const idIsValid = Types.ObjectId.isValid(id);
+    if (!idIsValid) {
+      throw new BadRequestException("Invalid id !");
+    }
     const todo = await this.todoModel.deleteOne({ _id: id }).exec();
     if (todo.deletedCount === 0) {
       throw new Error("Todo not found !");
